+++ data/build.gradle
@@ -89,6 +89,8 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:4.9.3'
     implementation 'org.jsoup:jsoup:1.13.1'
 
+    implementation "androidx.datastore:datastore-preferences:1.0.0"
+
 }
 
 kapt {
