+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/settings/PrefType.kt
@@ -1,0 +1,16 @@
+package net.pantasystem.milktea.data.infrastructure.settings
+
+sealed interface PrefType {
+    data class StrPref(
+        val value: String?
+    ) : PrefType
+
+    data class BoolPref(
+        val value: Boolean
+    ) : PrefType
+
+    data class IntPref(
+        val value: Int
+    ) : PrefType
+
+}
